Address review comments
- Moved sample GraphQL queries into their own files.

- Combined test-integrations and test-server-integration jobs as time savings are/will be
minimal even with a full suite of tests.

- Removed export of db client from the admin server.
