+++ package.json
@@ -8,8 +8,7 @@
     "start:dev": "npm run migrate:deploy && npm run build && npm run watch",
     "test-ci": "npm test",
     "test": "jest \"\\.spec\\.ts\"",
-    "test-integrations": "jest --testPathPattern=src/database --runInBand",
-    "test-server-integrations": "jest \"\\.server\\.integration\\.ts\" --runInBand",
+    "test-integrations": "jest \"\\.integration\\.ts\" --runInBand",
     "lint-check": "eslint --fix-dry-run \"src/**/*.ts\"",
     "lint-fix": "eslint --fix \"src/**/*.ts\"",
     "db:generate-client": "prisma generate",
