+++ src/test/admin-server/mutations/mutations.gql.ts
@@ -1,0 +1,79 @@
+import { gql } from 'apollo-server';
+
+/**
+ * Sample mutations for Apollo Server integration tests as used in
+ * Curation Admin Tools Frontend
+ */
+export const CREATE_COLLECTION_AUTHOR = gql`
+  mutation createCollectionAuthor(
+    $name: String!
+    $slug: String
+    $bio: Markdown
+    $imageUrl: Url
+    $active: Boolean
+  ) {
+    createCollectionAuthor(
+      data: {
+        name: $name
+        slug: $slug
+        bio: $bio
+        imageUrl: $imageUrl
+        active: $active
+      }
+    ) {
+      externalId
+      name
+      slug
+      bio
+      imageUrl
+      active
+    }
+  }
+`;
+
+export const UPDATE_COLLECTION_AUTHOR = gql`
+  mutation updateCollectionAuthor(
+    $externalId: String!
+    $name: String!
+    $slug: String!
+    $bio: Markdown
+    $imageUrl: Url
+    $active: Boolean
+  ) {
+    updateCollectionAuthor(
+      data: {
+        externalId: $externalId
+        name: $name
+        slug: $slug
+        bio: $bio
+        imageUrl: $imageUrl
+        active: $active
+      }
+    ) {
+      externalId
+      name
+      slug
+      bio
+      imageUrl
+      active
+    }
+  }
+`;
+
+export const UPDATE_COLLECTION_AUTHOR_IMAGE_URL = gql`
+  mutation updateCollectionAuthorImageUrl(
+    $externalId: String!
+    $imageUrl: Url!
+  ) {
+    updateCollectionAuthorImageUrl(
+      data: { externalId: $externalId, imageUrl: $imageUrl }
+    ) {
+      externalId
+      name
+      slug
+      bio
+      imageUrl
+      active
+    }
+  }
+`;
